% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute_education.R
\name{.impute_education}
\alias{.impute_education}
\title{Impute education data for block groups with universe population 0}
\usage{
.impute_education(bg_df, tracts_df)
}
\arguments{
\item{bg_df}{A data frame with block group data}

\item{tracts_df}{A data frame with tract data}
}
\value{
A data frame with imputed education data
}
\description{
This function imputes the education data for block groups with universe population 0.
It should be performed on the block group data frame after NSES has been calculated.
}
\keyword{internal}
