% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nses_master.R
\name{get_nses}
\alias{get_nses}
\title{Calculate NSES index and quartiles for all block groups in all 50 states and Washington, D.C.}
\usage{
get_nses(
  bg_df,
  tracts_df,
  columns = c("state", "tract_id", "bg_id", "geo_id", "total_total_population")
)
}
\arguments{
\item{bg_df}{A data frame of US Census block group data}

\item{tracts_df}{A data frame of US Census census tract data}

\item{columns}{A character vector of column names to include in the output data frame, along with NSES columns}
}
\value{
A data frame of US Census block group data with NSES index and quartiles
}
\description{
This function calculates the NSES index and quartiles for all block groups in all 50 states and Washington, D.C.
It takes in two data frames of US Census Data: one with block group data and one with tract data.
The function imputes missing values for block groups with values from the corresponding census tract.
}
