% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nses_constants.R
\docType{data}
\name{nses_concepts}
\alias{nses_concepts}
\title{NSES concepts to filter Census variables by when importing data}
\format{
An object of class \code{character} of length 6.
}
\usage{
nses_concepts
}
\description{
This is a list of regular expressions that can be used to filter Census variables to obtain
data required to comput the NSES index.
}
\keyword{datasets}
