% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nses_constants.R
\docType{data}
\name{nses_states}
\alias{nses_states}
\title{NSES states}
\format{
An object of class \code{data.frame} with 51 rows and 3 columns.
}
\usage{
nses_states
}
\description{
This is a data frame containing information about the all states in the United States
and Washington, D.C. It is used to filter out territories and other non-state entities.
}
\keyword{datasets}
